+++ src/CHANGES.rst
@@ -14,7 +14,26 @@
 CHANGELOG
 *********
 
-2.0.8 (2018-07-06)
+2.1 (2018-08-31)
+================
+
+Features
+--------
+
+- Added offset+limit case to function ``assert_container_listing``.
+- Added fix for memory leaks on pyramid segment cache.
+
+Changes
+-------
+
+- ``WebApp.request`` and ``WebApp.root`` has marked as deprecated.
+
+Bug Fixes
+---------
+
+- Fixed testing result headers inside of ``assert_resource()`` function.
+
+2.0.6 (2018-07-06)
 ==================
 
 Bug Fixes